Add --splitjumps mode (#2)
I'm working on a Lua script to track whether a player has seen all possible text in the game; it checks based on reads to CCS text block start addresses.  But because some blocks have conditional branches in their middles, the non-branching option at the ends of those blocks weren't being acknowledged as something unseen.   Using the new -s flag that I've added here will produce a dump where a new block is started whenever a conditional jump control code is encountered, so that "blocks" (in this mode) are only ever a chunk of text that's guaranteed to be printed in its entirety.